chore: move MentionList related classes to separate files
