+++ src/dotnet/Chat.UI.Blazor/Components/MentionList/IMentionsRetriever.cs
@@ -1,0 +1,6 @@
+namespace ActualChat.Chat.UI.Blazor.Components;
+
+public interface IMentionsRetriever
+{
+    Task<IEnumerable<Mention>> GetMentions(string search, int limit, CancellationToken cancellationToken);
+}
