+++ src/dotnet/Chat.UI.Blazor/Components/MentionList/Mention.cs
@@ -1,0 +1,3 @@
+namespace ActualChat.Chat.UI.Blazor.Components;
+
+public record Mention(string Id, string Name);
